+++ src/shared/main/tv/codely/shared/domain/DecimalValueObject.java
@@ -3,11 +3,11 @@ package tv.codely.shared.domain;
 import java.util.Objects;
 
 public abstract class DecimalValueObject {
-    private float value;
+    private double value;
 
-    public DecimalValueObject(float value){ this.value = value; }
+    public DecimalValueObject(double value){ this.value = value; }
 
-    public float value() { return value; }
+    public double value() { return value; }
 
     @Override
     public boolean equals(Object o) {
